requirements.txt
sqlalchemy>=2.0
pydantic>=2.9
python-dotenv>=1.0
pandas>=2.0
openai>=1.0
requests>=2.0
openpyxl>=3.1
